pyproject.toml
[project]
name = "llm-evals-notes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firecrawl-py>=2.16.5",
    "python-dotenv>=1.1.1",
    "python-slugify>=8.0.4",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]
